+++ Compression/Zlib/Tree.cs
@@ -267,7 +267,7 @@ namespace PMDCP.Compression.Zlib
                         continue;
                     if (tree[m * 2 + 1] != bits)
                     {
-                        s.opt_len = (int) (s.opt_len + ((long) bits - (long) tree[m * 2 + 1]) * (long) tree[m * 2]);
+                        s.opt_len = (int) (s.opt_len + (bits - (long) tree[m * 2 + 1]) * tree[m * 2]);
                         tree[m * 2 + 1] = (short) bits;
                     }
                     n--;
